Moved mkrepo script into its own repository under GPLv3+
This simple script has been copied verbatim (with the exception of the addition
of a license header); it was created to encourage developers---who may not have
much experience with GNU/Linux, Git, or the boxes on which the repositories are
hosted---to create well-factored repositories instead of monolithic ones.

This script is very young and, as such, does not do much. It is also
LoVullo-specific in some regards; this will change.